ott-admin/backend/location_service/requirements.txt
fastapi==0.71.0
pydantic==1.9.0
uvicorn==0.16.0
redis-py-cluster==2.1.3